refactor: create new folder structure for modular architecture
